Moved menu invalidation to the helper
